Move AddPromptsBookmarkModal to common components
