+++ src/js/popup/components/PromptsBookmark/PromptsBookmarkList/index.vue
@@ -48,7 +48,7 @@
  
 <script>
 
-import AddPromptsBookmarkModal from "./AddPromptsBookmarkModal/index.vue";
+import AddPromptsBookmarkModal from "~/js/popup/components/common/AddPromptsBookmarkModal/index.vue";
 import PromptsBookmarkItem from "./PromptsBookmarkItem/index.vue";
 import { useEnhancer } from "./enhancer";
