+++ src/js/popup/components/common/AddPromptsBookmarkModal/style.scss
